refactor: change MainActivity name and build a simple UI enhance androidX use in gradle.properties
