+++ app/src/main/AndroidManifest.xml
@@ -39,7 +39,7 @@
                 android:value="" />
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".HomeActivity"
             android:exported="true"
             android:theme="@style/Theme.Home">
