+++ app/src/main/res/drawable/ic_big_gear.xml
@@ -1,0 +1,5 @@
+<vector android:autoMirrored="true" android:height="512dp"
+    android:viewportHeight="512" android:viewportWidth="512"
+    android:width="512dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#000" android:pathData="M438.58,296.07l53.3,-14.69 -0.02,-50.73 -53.28,-14.74c-4.76,-21.9 -13.44,-42.33 -25.18,-60.61l27.32,-48.16 -35.9,-35.85 -48.08,27.3C338.4,86.82 317.98,78.16 296.05,73.35l-14.75,-53.22L230.64,20.12l-14.77,53.22c-21.87,4.81 -42.32,13.47 -60.61,25.24L107.13,71.28l-35.88,35.88 27.3,48.08c-11.74,18.34 -20.4,38.76 -25.24,60.66l-53.2,14.74v50.73l53.2,14.69a186.28,186.28 0,0 0,25.24 60.66l-27.3,48.16 35.88,35.83 48.13,-27.27a186.22,186.22 0,0 0,60.66 25.19l14.72,53.25 50.69,-0.02 14.72,-53.22c21.92,-4.81 42.35,-13.47 60.69,-25.21l48.13,27.33 35.81,-35.88 -27.27,-48.15c11.71,-18.31 20.4,-38.71 25.18,-60.63zM200.59,122.39h110.82l78.33,78.36v9.21h-55.16c-15.84,-26.93 -45.13,-45.05 -78.58,-45.05 -33.44,0 -62.74,18.12 -78.58,45.05L122.23,209.96v-9.21zM256,210.94c24.88,0 45.07,20.2 45.07,45.05 0,24.93 -20.19,45.1 -45.07,45.1 -24.87,0 -45.07,-20.17 -45.07,-45.1 0,-24.86 20.2,-45.05 45.07,-45.05zM311.41,389.99L200.59,389.99l-78.36,-78.36v-9.61h55.11a91.66,91.66 0,0 0,14.25 18.42c17.21,17.21 40.08,26.68 64.41,26.68 24.33,0 47.21,-9.48 64.42,-26.68a91.64,91.64 0,0 0,14.25 -18.42h55.08v9.61z"/>
+</vector>
