+++ app/src/main/res/drawable/ic_flame.xml
@@ -1,0 +1,5 @@
+<vector android:autoMirrored="true" android:height="512dp"
+    android:viewportHeight="512" android:viewportWidth="512"
+    android:width="512dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#000" android:pathData="M245.05,15.51c34.29,48.81 -23.53,320.54 -90.3,136.72C106.8,325.11 38.96,332.52 38.88,252.55c-71.6,79.31 43.82,220.77 87.38,243.93h52.13c-45.92,-40.02 -76.78,-78 -82.18,-135.97 47.31,9.42 71.86,20.96 81.26,-62.05 60.74,86.59 100.94,-49.38 137.18,-107.12 -1.65,40.32 -3.34,93.46 22.85,129.89 8.74,12.14 33.23,16.11 54.74,15.81 -9.92,16.08 -44.85,69.38 -17.01,89.2 27.84,19.82 33.07,-0.38 25.86,16.18 -13.26,20.88 -22.99,39.38 -59.07,54.06h56.06c59.44,-18.72 111.81,-91.66 94.61,-135.54 -22.01,18.66 -43.77,30.9 -61.29,29.54 49.12,-72.08 37.84,-145.9 14.75,-221.34 -20.22,72.38 -33.49,82.5 -54.58,99.52 29.1,-68.66 -85.44,-214.45 -146.51,-253.15z"/>
+</vector>
