+++ app/src/main/res/drawable/ic_rolling_suitcase.xml
@@ -1,0 +1,5 @@
+<vector android:autoMirrored="true" android:height="512dp"
+    android:viewportHeight="512" android:viewportWidth="512"
+    android:width="512dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#000" android:pathData="M256,18.76c-19.2,0 -36,3.4 -42.5,4.92h-0.1c-7.2,1.84 -12.2,8.26 -12.2,15.62v22.78h25v-15.5c7.1,-1.32 17.9,-2.82 29.8,-2.82 11.9,0 22.6,1.49 29.8,2.83v15.52h25L310.8,39.33c0,-7.36 -4.9,-13.79 -12.1,-15.6h-0.1c-6.6,-1.57 -23.4,-4.97 -42.6,-4.97zM205.3,75.7L205.3,172L222,172v-13.5c5.3,-3 17.1,-8.4 34,-8.4s28.7,5.4 34,8.4L290,172h16.7L306.7,75.7L290,75.7L290,140c-8.1,-3.3 -19.6,-6.5 -34,-6.5 -14.4,0 -25.9,3.2 -34,6.5L222,75.7h-16.7zM183,184.4c-22.2,0 -40.2,18 -40.2,40.2L142.8,417c0,7 1.9,14 5.3,20h44.3c6.7,0 13.1,2 17.8,7 3.6,4 6,8 7,13h77.6c1,-5 3.4,-9 7,-13 4.7,-5 11.1,-7 17.8,-7h44.3c3.4,-6 5.3,-13 5.3,-20L369.2,224.6c0,-22.2 -18,-40.2 -40.2,-40.2L183,184.4zM196,202.1h120c9.2,0 18.1,3.6 24.6,10.1s10.1,15.4 10.1,24.5v30.2c0,4.6 -3.7,8.3 -8.3,8.3h-145c-4.6,0 -8.3,-3.7 -8.3,-8.3 0,-4.6 3.7,-8.4 8.3,-8.4L334,258.5v-21.8c0,-4.7 -1.9,-9.3 -5.2,-12.7 -3.4,-3.4 -8,-5.3 -12.8,-5.3L196,218.7c-4.8,0 -9.4,1.9 -12.8,5.3 -3.3,3.4 -5.2,8 -5.2,12.7v30.2c0,4.6 -3.8,8.3 -8.4,8.3 -4.6,0 -8.3,-3.7 -8.3,-8.3v-30.2c0,-9.1 3.6,-18 10.1,-24.5s15.4,-10.1 24.6,-10.1zM169.6,293.9c4.6,0 8.4,3.7 8.4,8.3v91.3c0,4.8 1.9,9.5 5.2,12.5 3.4,4 8,6 12.8,6h120c4.8,0 9.4,-2 12.8,-6 3.3,-3 5.2,-7.7 5.2,-12.5v-83L197.4,310.5c-4.6,0 -8.3,-3.7 -8.3,-8.3 0,-4.6 3.7,-8.3 8.3,-8.3h145c4.6,0 8.3,3.7 8.3,8.3v91.3c0,9.5 -3.6,18.5 -10.1,24.5 -6.5,7 -15.4,10 -24.6,10L196,428c-9.2,0 -18.1,-3 -24.6,-10 -6.5,-6 -10.1,-15 -10.1,-24.5v-91.3c0,-4.6 3.7,-8.3 8.3,-8.3zM184,452.8c-11.4,0 -20.6,9.2 -20.6,20.6 0,11.3 9.2,20.5 20.6,20.5 11.3,0 20.6,-9.2 20.6,-20.5 0,-11.4 -9.3,-20.6 -20.6,-20.6zM328.2,452.8c-11.4,0 -20.6,9.2 -20.6,20.6 0,11.3 9.2,20.5 20.6,20.5 11.3,0 20.6,-9.2 20.6,-20.5 0,-11.4 -9.3,-20.6 -20.6,-20.6zM184,465.3c4.4,0 8.1,3.6 8.1,8.1 0,4.4 -3.7,8 -8.1,8 -4.5,0 -8,-3.6 -8,-8 0,-4.5 3.5,-8.1 8,-8.1zM328.2,465.3c4.4,0 8.1,3.6 8.1,8.1 0,4.4 -3.7,8 -8.1,8 -4.5,0 -8,-3.6 -8,-8 0,-4.5 3.5,-8.1 8,-8.1z"/>
+</vector>
