--- app/src/main/res/layout/activity_main.xml
@@ -1,94 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:padding="16dp"
-    android:gravity="center_horizontal"
-    android:background="@drawable/bluried"
-    tools:context=".MainActivity">
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="14dp"
-        android:text="@string/easter_eggs"
-        android:fontFamily="serif-monospace"
-        android:textStyle="bold"
-        android:textSize="48sp"
-        android:textColor="@color/Beige"
-        android:textAlignment="center" />
-
-    <TextView
-        android:id="@+id/intro"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/intro"
-        android:fontFamily="serif-monospace"
-        android:textStyle="bold|italic"
-        android:textSize="28sp"
-        android:textColor="@color/Beige"
-        android:textAlignment="center"
-        android:layout_marginTop="60dp"
-        />
-
-    <Button
-        android:id="@+id/brew_button"
-        android:layout_width="300dp"
-        android:layout_height="56dp"
-        android:layout_marginTop="24dp"
-        android:text="@string/brew_a_person"
-        android:textColor="@color/white"
-        android:textStyle="bold"
-        android:textSize="16sp"
-        />
-
-    <TextView
-        android:id="@+id/counter"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="30sp"
-        android:text="@string/people_count"
-        android:textColor="@color/Ash_Grey"
-        android:textSize="78sp"
-        android:fontFamily="casual"
-        android:textStyle="bold"
-        />
-
-    <Button
-        android:id="@+id/kill_button"
-        android:layout_width="300dp"
-        android:layout_height="56dp"
-        android:layout_marginTop="24dp"
-        android:backgroundTint="@color/Ruber"
-        android:text="@string/kill_a_person"
-        android:textColor="@color/white"
-        android:textSize="16sp"
-        android:textStyle="bold"
-        android:visibility="invisible"
-        tools:visibility="visible" />
-
-    <Button
-        android:id="@+id/hellGateBtn"
-        android:layout_width="wrap_content"
-        android:layout_height="50dp"
-        android:layout_marginTop="34dp"
-        android:backgroundTint="@color/Sizzling_Red"
-        android:text="@string/hellGateIntro"
-        android:textColor="@color/white"
-        android:textSize="14sp"
-        android:textStyle="bold"
-        android:visibility="invisible"
-        tools:visibility="visible" />
-
-    <nl.dionsegijn.steppertouch.StepperTouch
-        android:id="@+id/stepperTouch"
-        android:layout_width="160dp"
-        android:layout_height="60dp"
-        android:layout_marginTop="30dp"
-        app:stepperActionsColor="@color/white"
-        />
-
-</LinearLayout>
