+++ gradle.properties
@@ -22,3 +22,4 @@ kotlin.code.style=official
 # thereby reducing the size of the R class for that library
 android.nonTransitiveRClass=true
 org.gradle.unsafe.configuration-cache=true
+android.enableJetifier=true
